split an awkwardly big test
